web/requirements.txt
asgiref==3.6.0
azure-identity==1.16.1
certifi==2023.7.22
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Django==4.1.13
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
django-model-utils==4.3.1
django-rest-swagger==2.2.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-csv==2.1.1
djangorestframework-gis==1.0
djangorestframework-xml==2.0.0
drf-amsterdam==0.4.1
drf-extensions==0.7.1
drf-hal-json==0.10.0
drf-nested-fields==0.9.5
idna==3.4
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.2
openapi-codec==1.3.2
Pillow==10.3.0
psycopg2==2.9.5
python-swiftclient==4.1.0
pytz==2022.7.1
requests==2.32.0
sentry-sdk==1.14.0
simplejson==3.18.1
six==1.16.0
sqlparse==0.5.0
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==1.26.19
